Simplifying methods and moving from retrievals to utils
